feat: move badge editor to separate component
